Keep fixture graph cache independent of query gates
